reordered Nostr files and removed depricate files
Each state has his own nostr class to better understand it
Nostrstettings now hold only values. NostrHelper now has helper functions for Nostr Events and Requests
Probably added to much NostrSettings.players ... to add the names but it works now and I don't know wich I can delete and which not. It stays like this for a while
When a thrid one is joining the players are only correct by the host. Functional players sync is in progress..